Move extensions to own file
